+++ Platform/OpenCore/OpenCoreUefi.c
@@ -439,9 +439,9 @@ OcLoadUefiSupport (
 
   DEBUG ((
     DEBUG_INFO,
-    "OC: RBVR %d RBVF %d\n",
+    "OC: RBVR %d DDBR %d\n",
     Config->Uefi.Quirks.RequestBootVarRouting,
-    Config->Uefi.Quirks.RequestBootVarFallback
+    Config->Uefi.Quirks.DeduplicateBootOrder
     ));
 
   //
@@ -455,15 +455,7 @@ OcLoadUefiSupport (
     &Config->Uefi.Quirks.RequestBootVarRouting
     );
 
-  gRT->SetVariable (
-    OC_BOOT_FALLBACK_VARIABLE_NAME,
-    &gOcVendorVariableGuid,
-    OPEN_CORE_INT_NVRAM_ATTR,
-    sizeof (Config->Uefi.Quirks.RequestBootVarFallback),
-    &Config->Uefi.Quirks.RequestBootVarFallback
-    );
-
-  if (Config->Uefi.Quirks.RequestBootVarFallback) {
+  if (Config->Uefi.Quirks.DeduplicateBootOrder) {
     BootOrder = OcGetBootOrder (
       &gEfiGlobalVariableGuid,
       FALSE,